Must initialize an EMS context when retrieving parameters
Without this EMS immediately dumps error messages to standard output
even if the code subsequently annuls the error. This makes for very
confusing output from par_get when it fails to open a v4 parameter
file and then opens it in v5 mode. emsBegin/emsEnd are needed to
properly capture the error state.